# lib/api-zod/src/generated/types/pricingOption.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PricingOption {
  duration: string;
  /** @minimum 0 */
  price: number;
  /** @nullable */
  salePrice?: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  priceUsd?: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  salePriceUsd?: number | null;
}